Taxons without rank should still get reported
The current code for the DBSeq class removes any taxonomic name that
has a rank of 'no rank'. This can change the returned taxonomy of the
sequence significantly and is incorrect with what is reported at NCBI. It
is further confusing as the taxonomic names are stored correctly in the
database but are filtered out after their retieval.